+++ test/vlog/pp6.v
@@ -1,0 +1,5 @@
+one
+`include "pp6.vh" // Comment
+two
+`include "foo" 123
+`include "not.here" // Test